[REF] refactor architecture module create a generic backend for all nosql backend
