+++ connector_nosql_solr/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from . import backend
+from . import models
+from . import unit
