+++ connector_nosql_solr/tests/__init__.py
