+++ connector_nosql_solr/unit/__init__.py
