+++ connector_nosql_solr/unit/mapper.py
