+++ connector_nosql_solr_product/__init__.py
@@ -2,5 +2,3 @@
 # © 2016 Akretion (http://www.akretion.com)
 # Sébastien BEAU <sebastien.beau@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-from . import models
