+++ connector_nosql_solr_product/__openerp__.py
@@ -3,14 +3,15 @@
 # Sébastien BEAU <sebastien.beau@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-{'name': 'Connector Search Engine',
+{'name': 'SolR Catalog Indexer',
  'version': '0.0.1',
  'author': 'Akretion',
  'website': 'www.akretion.com',
  'license': 'AGPL-3',
  'category': 'Generic Modules',
  'depends': [
-     'connector',
+     'connector_nosql_solr',
+     'connector_nosql_product',
  ],
  'data': [
  ],
