+++ connector_nosql_solr_product/tests/__init__.py
