+++ connector_nosql_solr_product/tests/data_export.py
@@ -6,7 +6,7 @@
 
 export_product_3_4 = [('add',
   2,
-  [{'backend_id/id_is': 2,
+  [{'backend_id/id_is': 1,
     'backend_id/name_ss': u'Demo Solr Backend',
     'create_uid/id_is': 1,
     'create_uid/name_ss': u'Administrator',
@@ -15,10 +15,10 @@ export_product_3_4 = [('add',
     'sync_state_k_ss': u'to_update',
     'sync_state_ss': 'To update',
     'text': None,
-    'type': 'SolrProductTemplate',
+    'type': 'NosqlProductTemplate',
     'write_uid/id_is': 1,
     'write_uid/name_ss': u'Administrator'},
-   {'backend_id/id_is': 2,
+   {'backend_id/id_is': 1,
     'backend_id/name_ss': u'Demo Solr Backend',
     'create_uid/id_is': 1,
     'create_uid/name_ss': u'Administrator',
@@ -27,7 +27,7 @@ export_product_3_4 = [('add',
     'sync_state_k_ss': u'to_update',
     'sync_state_ss': 'To update',
     'text': None,
-    'type': 'SolrProductTemplate',
+    'type': 'NosqlProductTemplate',
     'write_uid/id_is': 1,
     'write_uid/name_ss': u'Administrator'}]),
  'commit']
