--- connector_search_engine/unit/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-# © 2016 Akretion (http://www.akretion.com)
-# Sébastien BEAU <sebastien.beau@akretion.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-
-from . import exporter
-from . import adapter
