--- solerp/__init__.py
@@ -1,3 +1,0 @@
-import connector
-import backend
-import solr_model
