--- solerp/security/ir.model.access.csv
@@ -1,2 +1,0 @@
-"id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
-"access_solr_backend","solr_backend connector manager","model_solr_backend","connector.group_connector_manager",1,1,1,1
