--- solerp/solr_menu.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-  <data>
-
-    <menuitem id="menu_solr_root"
-        parent="connector.menu_connector_root"
-        name="Solr"
-        sequence="10"
-        groups="connector.group_connector_manager"/>
-
-    <menuitem id="menu_solr_backend"
-        name="Backends"
-        parent="menu_solr_root"
-        action="action_solr_backend"/>
-
-  </data>
-</openerp>
