--- solerp/solr_model_view.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="view_solr_backend_form" model="ir.ui.view">
-            <field name="name">solr.backend.form</field>
-            <field name="model">solr.backend</field>
-            <field name="arch" type="xml">
-                <form string="solr Backend" version="7.0">
-                    <header>
-                      <button name="index_all"
-                        type="object"
-                        class="oe_highlight"
-                        string="Index All"/>
-                    </header>
-                    <sheet>
-                        <label for="name" class="oe_edit_only"/>
-                        <h1>
-                            <field name="name" class="oe_inline" />
-                        </h1>
-                        <group name="solr" string="solr Configuration">
-                            <group colspan="4" col="4">
-                                <field name="version" colspan="4"/>
-                                <field name="location" placeholder="e.g. http://my.solr.com"
-                                    colspan="2"/>
-                                <field name="default_lang_id" colspan="2"/>
-                                <field name="username" colspan="2"/>
-                                <field name="password" colspan="2"/>
-                            </group>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-
-        <record id="view_solr_backend_tree" model="ir.ui.view">
-            <field name="name">solr.backend.tree</field>
-            <field name="model">solr.backend</field>
-            <field name="arch" type="xml">
-                <tree string="solr Backend" version="7.0">
-                    <field name="name"/>
-                    <field name="location"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="action_solr_backend" model="ir.actions.act_window">
-            <field name="name">solr Backends</field>
-            <field name="res_model">solr.backend</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-            <field name="view_id" ref="view_solr_backend_tree"/>
-        </record>
-
-        <record id="view_solr_exports_form" model="ir.ui.view">
-            <field name="name">solr.exports.form</field>
-            <field name="model">ir.exports</field>
-            <field name="arch" type="xml">
-                <form string="Export" version="7.0">
-                    <sheet>
-                        <label for="name" class="oe_edit_only"/>
-                        <h1>
-                            <field name="name" class="oe_inline" />
-                        </h1>
-                        <group name="solr" string="Export Configuration">
-                            <group colspan="4" col="4">
-                                <field name="resource" colspan="4"/>
-                                <field name="export_fields" colspan="4"/>
-                            </group>
-                        </group>
-                    </sheet>
-                </form>
-            </field>
-        </record>
-
-    </data>
-</openerp>
