--- solerp_catalog/datas/cron_data.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data noupdate="1">
-<record forcecreate="True" id="ir_cron_import_sale_orders" model="ir.cron">
-    <field name="name">SolR Export Product</field>
-    <field eval="False" name="active"/>
-    <field name="user_id" ref="base.user_root"/>
-    <field name="interval_number">1</field>
-    <field name="interval_type">days</field>
-    <field name="numbercall">-1</field>
-    <field eval="False" name="doall"/>
-    <field eval="'solr.backend'" name="model"/>
-    <field eval="'_scheduler_export_product'" name="function"/>
-    <field eval="'()'" name="args"/>
-</record> 
-    </data>
-</openerp>
