--- solerp_catalog/views/product_view.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-
-<record id="view_product_template_form" model="ir.ui.view">
-    <field name="model">product.template</field>
-    <field name="inherit_id" ref="connector_base_product.product_template_common_form"/>
-    <field name="arch" type="xml">
-        <page name="connector" position="attributes">
-            <attribute name="invisible">0</attribute>
-        </page>
-        <page name="connector" position="inside">
-            <field name="solr_bind_ids">
-                <tree editable="bottom">
-                    <field name="backend_id"/>
-                    <field name="sync_state"/>
-                    <field name="date_modified"/>
-                    <field name="date_syncronized"/>
-                </tree>
-            </field>
-        </page>
-    </field>
-</record>
-
-    </data>
-</openerp>
